Invalid function definition fix

+++ src/lib/ndpi_main.c
@@ -6574,7 +6574,7 @@ void NDPI_DUMP_BITMASK(NDPI_PROTOCOL_BITMASK a) {
 }
 #endif
 
-u_int16_t ndpi_get_api_version() {
+int ndpi_get_api_version() {
   return(NDPI_API_VERSION);
 }
